Refactor OrderConfirmationModal and TradingModal utility to improve type management and clarity
- Updated import paths for TriggerCondition to streamline type usage in OrderConfirmationModal and TradingModal.util.
- Renamed tokenBalances to walletTokenBalances in useDepositWithdrawData for better clarity and consistency across the codebase.
- Adjusted related references to ensure accurate data handling and improved maintainability.
